+++ app.js
@@ -27,6 +27,7 @@ const user = require("./routes/user");
 const reset = require("./routes/reset");
 const debt = require('./routes/debt_reminder');
 const businessCards = require("./routes/businessCardRoute");
+const updates = require("./routes/update_routes");
 // const phone_call_api = require("./controllers/phone_call_api");
 app.use(cors());
 app.use(expressValidator());
@@ -70,6 +71,7 @@ app.use(reset);
 app.use(emailAPI);
 app.use(transactions);
 app.use(businessCards);
+app.use("/update", updates);
 app.use(store);
 app.use(google);
 app.use(facebook);
